+++ fabric/src/main/java/net/modgarden/barricade/fabric/BarricadeFabric.java
@@ -8,6 +8,7 @@ import net.fabricmc.fabric.api.event.registry.DynamicRegistries;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerConfigurationConnectionEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerConfigurationNetworking;
+import net.fabricmc.fabric.api.networking.v1.ServerLoginConnectionEvents;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
 import net.fabricmc.fabric.api.resource.ResourcePackActivationType;
 import net.fabricmc.loader.api.FabricLoader;
@@ -38,7 +39,7 @@ public class BarricadeFabric implements ModInitializer {
                 return;
             ServerConfigurationNetworking.send(handler, new SetServerContextClientboundPacket());
         });
-        ServerLifecycleEvents.SERVER_STARTING.register(minecraftServer -> Barricade.setServerContext());
+        ServerLifecycleEvents.SERVER_STARTING.register(minecraftServer -> Barricade.serverContext = true);
 
         FabricLoader.getInstance().getModContainer(Barricade.MOD_ID).ifPresent(modContainer -> {
             ResourceManagerHelper.registerBuiltinResourcePack(Barricade.asResource("modded_rendering"), modContainer, Component.translatable("resourcePack.barricade.modded_rendering.name"), ResourcePackActivationType.DEFAULT_ENABLED);
@@ -47,7 +48,7 @@ public class BarricadeFabric implements ModInitializer {
         DynamicRegistries.registerSynced(BarricadeRegistries.ADVANCED_BARRIER, AdvancedBarrier.DIRECT_CODEC);
 
         ItemGroupEvents.modifyEntriesEvent(CreativeModeTabs.OP_BLOCKS).register(entries -> {
-            if (!entries.shouldShowOpRestrictedItems())
+            if (!entries.shouldShowOpRestrictedItems() || Barricade.serverContext)
                 return;
             entries.addAfter(Items.BARRIER,
                     BarricadeItems.UP_BARRIER,